Add functional display
Added a functional display for the calculator. When nothing has been
entered, the display shows "0". When the user enters the first number,
the display shows the first number. When the user clicks one of the
operators and then enters the second number, the second number is
displayed. When the user clicks the equals button, the result is
displayed. Numbers greater than one digit are now supported.