Allow multiple operations
Changed the operator button behaviour so the user can carry out an
operation on more than two numbers without clicking the equals button.
The initial operation is carried out and the result is used as the first
number in the next operation. As outlined in the specification, this
does not preserve order of operations. The user can also carry out
operations on the result obtained after clicking the equals button.