Replace rounding with exponential notation
Modified the add, subtract, multiply and divide functions so that if a
number is too large to display, it is displayed in exponential notation.
Further, the number of number characters the user can type in has been
limited to 8 (leaving room for a decimal point and negation if they
choose).